Refractoring Main Page Code
